# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees/{id}/vacations:
    get:
      operationId: listEmployeeVacations
      tags: [vacations]
      summary: List vacations for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of vacations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListVacationsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createVacation
      tags: [vacations]
      summary: Register a vacation period for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVacationBody"
      responses:
        "201":
          description: Vacation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacationItem"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /vacations/{id}:
    delete:
      operationId: deleteVacation
      tags: [vacations]
      summary: Delete a vacation period
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Vacation deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /vacations/{id}/status:
    put:
      operationId: updateVacationStatus
      tags: [vacations]
      summary: Approve or reject a vacation request (managers only)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVacationStatusBody'
      responses:
        '200':
          description: Vacation status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacationItem'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - manager role required
        '404':
          description: Vacation not found
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    VacationItem:
      type: object
      required: [id, employeeId, startDate, endDate, durationDays, notes, status, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        durationDays:
          type: integer
        notes:
          type: [string, 'null']
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
    ListVacationsResponse:
      type: object
      required: [vacations]
      properties:
        vacations:
          type: array
          items:
            $ref: "#/components/schemas/VacationItem"
    UpdateVacationStatusBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
    CreateVacationBody:
      type: object
      required: [startDate, endDate]
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: [string, 'null']
...