# lib/api-zod/src/generated/types/updateVacationStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateVacationStatusBodyStatus } from "./updateVacationStatusBodyStatus";

export interface UpdateVacationStatusBody {
  status: UpdateVacationStatusBodyStatus;
}